Move `BibPage` to `pages` folder
